+++ docker/types/services.py
@@ -20,7 +20,9 @@ class TaskTemplate(dict):
           individual container created as part of the service.
         restart_policy (RestartPolicy): Specification for the restart policy
           which applies to containers created as part of this service.
-        placement (:py:class:`list`): A list of constraints.
+        placement (Placement): Placement instructions for the scheduler.
+            If a list is passed instead, it is assumed to be a list of
+            constraints as part of a :py:class:`Placement` object.
         force_update (int): A counter that triggers an update even if no
             relevant parameters have been changed.
     """
@@ -33,7 +35,7 @@ class TaskTemplate(dict):
             self['RestartPolicy'] = restart_policy
         if placement:
             if isinstance(placement, list):
-                placement = {'Constraints': placement}
+                placement = Placement(constraints=placement)
             self['Placement'] = placement
         if log_driver:
             self['LogDriver'] = log_driver
@@ -107,11 +109,14 @@ class ContainerSpec(dict):
         if labels is not None:
             self['Labels'] = labels
         if mounts is not None:
+            parsed_mounts = []
             for mount in mounts:
                 if isinstance(mount, six.string_types):
-                    mounts.append(Mount.parse_mount_string(mount))
-                    mounts.remove(mount)
-            self['Mounts'] = mounts
+                    parsed_mounts.append(Mount.parse_mount_string(mount))
+                else:
+                    # If mount already parsed
+                    parsed_mounts.append(mount)
+            self['Mounts'] = parsed_mounts
         if stop_grace_period is not None:
             self['StopGracePeriod'] = stop_grace_period
 
@@ -124,7 +129,7 @@ class ContainerSpec(dict):
 class Mount(dict):
     """
     Describes a mounted folder's configuration inside a container. A list of
-    :py:class:`Mount`s would be used as part of a
+    :py:class:`Mount` would be used as part of a
     :py:class:`~docker.types.ContainerSpec`.
 
     Args:
@@ -302,8 +307,8 @@ class RestartPolicy(dict):
         condition (string): Condition for restart (``none``, ``on-failure``,
           or ``any``). Default: `none`.
         delay (int): Delay between restart attempts. Default: 0
-        attempts (int): Maximum attempts to restart a given container before
-          giving up. Default value is 0, which is ignored.
+        max_attempts (int): Maximum attempts to restart a given container
+          before giving up. Default value is 0, which is ignored.
         window (int): Time window used to evaluate the restart policy. Default
           value is 0, which is unbounded.
     """
@@ -407,7 +412,7 @@ class ServiceMode(dict):
                 'replicas can only be used for replicated mode'
             )
         self[mode] = {}
-        if replicas:
+        if replicas is not None:
             self[mode]['Replicas'] = replicas
 
     @property
@@ -449,3 +454,28 @@ class SecretReference(dict):
             'GID': gid or '0',
             'Mode': mode
         }
+
+
+class Placement(dict):
+    """
+        Placement constraints to be used as part of a :py:class:`TaskTemplate`
+
+        Args:
+            constraints (list): A list of constraints
+            preferences (list): Preferences provide a way to make the
+                scheduler aware of factors such as topology. They are provided
+                in order from highest to lowest precedence.
+            platforms (list): A list of platforms expressed as ``(arch, os)``
+                tuples
+    """
+    def __init__(self, constraints=None, preferences=None, platforms=None):
+        if constraints is not None:
+            self['Constraints'] = constraints
+        if preferences is not None:
+            self['Preferences'] = preferences
+        if platforms:
+            self['Platforms'] = []
+            for plat in platforms:
+                self['Platforms'].append({
+                    'Architecture': plat[0], 'OS': plat[1]
+                })